requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
brainfuck-interpreter==1.0.0
charset-normalizer==2.1.1
colorlog==6.6.0
frozenlist==1.3.1
hikari==2.0.0.dev110
idna==3.3
multidict==6.0.2
Nuitka==1.0.6
ujson==5.4.0
uvloop==0.16.0
yarl==1.8.1
